requirements.txt
typer
pyyaml
Jinja2
pycryptodome